Initial pipeline process with name queues and worker processes
Also a simple driver to kick the process.

I still need to iron out a lot of crufts to make it more flexible in configurations
and not tightly coupling.

Also I need to solve the python path search problem for drivers and celery works
and make the whole system startup easier.