Fixed empty file response on local HPZ
The check to detect if we're running in production was introduced
to use `pyramid.response.FileResponse` in case when HPZ is running
locally. This fixes the problem that we're receiving file with valid
name, but invalid content (it's empty). Downloaded file is empty because
we're using `X-Sendfile` header to serve files (that is then handled by
Apache, which we don't have in local env). This might be removed when we
introduce Docker (or something similar) for running all components locally.