Rewrote `_partial_pdfunite` so it could work properly with >50 reports
There was an issue in `_partial_pdfunite` logic - it was skipping 50
reports. The fix was easy, but the existing code was unnecessarily
complex. Also, it wasn't handling situations when its output can
contain more than specified `file_limit`, which can lead to crashes (due
to file descriptors limit) if we have all of the workers working in
parallel by the same workflow with considerable amount of files, or having
one worker merging a lot of files at once (>50k).

DE1433