Added a filter on `dim_student.c.rec_status` into the join clause in the query at `_get_student_ids`
The filter on `rec_status = 'C'` was added in order to fix the crash,
that appeared when client was trying to download a PDF report for one
specific grade. This grade has 3 students that were downloaded with
typos in their names at first. Then those typos were fixed (presumably
by uploading another batch of data). Because of that, those students
have additional records in the `dim_student` table with
`rec_status == 'I'`. `_get_student_ids` function has a query, that was
not filtering on `rec_status = 'C'`, so few inactive records were
joined, and that's the reason for the crash (the actual crash happens in
`generate_isr_report_path_by_student_id`, when it is trying to compare
length of the `student_ids` and its local query results).

TA43239