Add additional manage command for some UDL and TSB init scripts
At this moment we have some init scripts (basically related to UDL and
TSB) which do not support to performing extra management commands,
for example: restart, status. Thus, when ansible plays book and tries to
restart such services it gets an error and that service remains off.
Another example - system was rebooted for some reason but it does not
have runlevel information for some services.

US38997/TA43266